+++ app/src/androidTest/java/com/github/se/gatherspot/run/model/EventUtilsTest.kt
@@ -3,12 +3,12 @@ package com.github.se.gatherspot.run.model
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.github.se.gatherspot.firebase.EventFirebaseConnection
-import com.github.se.gatherspot.model.EventUtils
 import com.github.se.gatherspot.model.Interests
 import com.github.se.gatherspot.model.Profile
 import com.github.se.gatherspot.model.event.Event
 import com.github.se.gatherspot.model.event.EventStatus
 import com.github.se.gatherspot.model.location.Location
+import com.github.se.gatherspot.model.utils.EventUtils
 import com.github.se.gatherspot.ui.eventUI.EventAction
 import java.time.LocalDate
 import java.time.LocalTime
