+++ app/src/androidTest/java/com/github/se/gatherspot/viewModel/EventUIViewModelTest.kt
@@ -6,12 +6,12 @@ import com.github.se.gatherspot.EnvironmentSetter.Companion.testLoginUID
 import com.github.se.gatherspot.firebase.EventFirebaseConnection
 import com.github.se.gatherspot.firebase.ProfileFirebaseConnection
 import com.github.se.gatherspot.firebase.RatingFirebaseConnection
-import com.github.se.gatherspot.model.EventUtils
 import com.github.se.gatherspot.model.Interests
 import com.github.se.gatherspot.model.Profile
 import com.github.se.gatherspot.model.Rating
 import com.github.se.gatherspot.model.event.Event
 import com.github.se.gatherspot.model.event.EventStatus
+import com.github.se.gatherspot.model.utils.EventUtils
 import com.github.se.gatherspot.ui.eventUI.EventRegistrationViewModel
 import com.github.se.gatherspot.ui.eventUI.EventUIViewModel
 import com.github.se.gatherspot.ui.eventUI.RegistrationState
