+++ app/src/main/java/com/github/se/gatherspot/ui/eventUI/CreateEvent.kt
@@ -2,7 +2,7 @@ package com.github.se.gatherspot.ui.eventUI
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.github.se.gatherspot.model.EventUtils
+import com.github.se.gatherspot.model.utils.EventUtils
 import com.github.se.gatherspot.ui.navigation.NavigationActions
 
 /**
